python/server/requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
opencv-python-headless==4.10.0.82
pillow==10.3.0
requests==2.32.3
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.2
tensorflow
scikit-learn